Move ILogger
