+++ extension/src/Logging/ConsoleLogger.ts
@@ -1,4 +1,5 @@
-import { ILogger, appendTimestamp } from "./LogHelper";
+import { ILogger } from "./ILogger";
+import { appendTimestamp } from "./LogHelper";
 
 export class ConsoleLogger implements ILogger {
   private static instance: ConsoleLogger;
