+++ extension/src/Logging/ILogger.ts
@@ -1,0 +1,4 @@
+export interface ILogger {
+  log(message?: string, ...optionalParams: string[]): void;
+  error(message?: string, ...optionalParams: string[]): void;
+}
