+++ extension/src/Logging/LogHelper.ts
@@ -1,16 +1,12 @@
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const timestamp = require("time-stamp");
+import { ILogger } from "./ILogger";
 import { NullLogger } from "./NullLogger";
 
 export let logger: ILogger = new NullLogger();
 
 export function setLogger(newLogger: ILogger): void {
   logger = newLogger;
-}
-
-export interface ILogger {
-  log(message?: string, ...optionalParams: string[]): void;
-  error(message?: string, ...optionalParams: string[]): void;
 }
 
 export function appendTimestamp(line?: string): string {
