+++ extension/src/Logging/NullLogger.ts
@@ -1,5 +1,5 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
-import { ILogger } from "./LogHelper";
+import { ILogger } from "./ILogger";
 
 export class NullLogger implements ILogger {
   log(_message?: string, ..._optionalParams: string[]): void {
