+++ extension/src/PowerShell.ts
@@ -1,5 +1,6 @@
 import * as Shell from "node-powershell";
-import { ILogger, logger } from "./Logging/LogHelper";
+import { ILogger } from "./Logging/ILogger";
+import { logger } from "./Logging/LogHelper";
 
 export class Powershell {
   private startTime: Date = new Date();
